+++ khoros/objects/__init__.py
@@ -2,17 +2,18 @@
 """
 :Module:            khoros.objects
 :Synopsis:          This module contains sub-modules that are specific to the various Community API objects.
-:Usage:             ``from khoros.objects import base``
-:Example:           ``node_id = base.get_node_id('https://community.khoros.com/t5/Khoros-Blog/bg-p/relnote', 'blog')``
+:Usage:             ``from khoros import objects``
+:Example:           ``objects.archives.archive(khoros_obj, '123', suggested_url, return_status=True)``
 :Created By:        Jeff Shurtliff
 :Last Modified:     Jeff Shurtliff
-:Modified Date:     11 May 2020
+:Modified Date:     07 Jun 2020
 """
 
-__all__ = ['albums', 'messages', 'roles', 'users']
+__all__ = ['albums', 'archives', 'messages', 'roles', 'users']
 
 # Import all submodules by default
 from . import albums
+from . import archives
 from . import messages
 from . import roles
 from . import users